pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TestMagick"
version = "0.1.0"
description = "Typst 기반 YAML/JSON 입력으로 문제지·정답지 PDF를 생성하는 로컬 CLI 도구."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Jinja2>=3.1.4,<4",
  "PyYAML>=6.0.1,<7",
  "pydantic>=2.6,<3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9",
  "ruff>=0.6.0,<1",
]

[project.scripts]
testmagick = "testmagick.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
testmagick = ["templates/*.j2"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
